fix: NextJS useSearchParams hook on server side
app-msa-92
