+++ app/(media)/tv-shows/TvShow.ts
@@ -1,4 +1,3 @@
-import { FALLBACK_ID_PREFIX } from "@/app/(media)/movies/Movie";
 import { Tmdb } from "@/models/services/Tmdb";
 import { Trakt } from "@/models/services/Trakt";
 import Media from "@/models/Media";
